backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
anthropic>=0.40
langgraph>=0.2
python-dotenv>=1.0
python-multipart>=0.0.9
aiofiles>=24.1
pymupdf>=1.24
python-docx>=1.1
openpyxl>=3.1
